+++ filament/src/details/SwapChain.h
@@ -63,6 +63,8 @@ public:
 
     void setFrameCompletedCallback(FrameCompletedCallback callback, void* user);
 
+    static bool isSRGBSwapChainSupported(FEngine& engine) noexcept;
+
 private:
     FEngine& mEngine;
     backend::Handle<backend::HwSwapChain> mSwapChain;